feat: delete powertag
