+++ src/features/manage-tags/index.tsx
@@ -1,17 +1,4 @@
 import {
-  DndContext,
-  DragEndEvent,
-  PointerSensor,
-  useSensor,
-  useSensors,
-} from '@dnd-kit/core'
-import {
-  arrayMove,
-  SortableContext,
-  verticalListSortingStrategy,
-} from '@dnd-kit/sortable'
-import { IconMenuOrder } from '@tabler/icons-react'
-import {
   Dispatch,
   SetStateAction,
   useCallback,
@@ -19,7 +6,7 @@ import {
   useState,
 } from 'react'
 
-import { SortableItem } from '../../components/SortableItem'
+import { TagProperties } from '../../components/TagProperties'
 import { Tag } from '..'
 
 export const ManageTags = ({
@@ -29,63 +16,23 @@ export const ManageTags = ({
   tags: Tag
   setTags: Dispatch<SetStateAction<Tag>>
 }) => {
-  const [localTags, setLocalTags] = useState<Tag>()
+  console.log(tags)
+  const [localTags, setLocalTags] = useState<Tag>({})
 
   useEffect(() => {
     setLocalTags(tags)
   }, [tags])
 
-  const sensors = useSensors(
-    useSensor(PointerSensor, {
-      activationConstraint: {
-        distance: 8,
-      },
-    }),
-  )
-
-  const handleDragEnd = useCallback(
-    (event: DragEndEvent) => {
-      const { active, over } = event
-
-      if (active.id !== over?.id) {
-        setLocalTags((prevTags) => {
-          if (!prevTags) return prevTags
-
-          const newTags = { ...prevTags }
-          const tagIndex = Object.keys(prevTags).find((key) =>
-            prevTags[key].some((prop) => prop.name === active.id),
-          )
-
-          if (!tagIndex) return prevTags
-
-          const activeIndex = prevTags[tagIndex].findIndex(
-            (prop) => prop.name === active.id,
-          )
-          const overIndex = prevTags[tagIndex].findIndex(
-            (prop) => prop.name === over?.id,
-          )
-
-          if (activeIndex === -1 || overIndex === -1) return prevTags
-
-          newTags[tagIndex] = arrayMove(
-            prevTags[tagIndex],
-            activeIndex,
-            overIndex,
-          )
-
-          console.log('New order of properties:', newTags[tagIndex])
-          return newTags
-        })
-      }
-    },
-    [], // Remove the tags dependency as we're using the function form of setLocalTags
-  )
-
-  const deleteProperty = useCallback(
-    (index: string, name: string) => {
-      console.log('HELLO WORLD')
+  const deletePowertag = useCallback(
+    (index: string) => {
       console.log(index)
-      console.log(name)
+      const currSavedTags = logseq.settings!.savedTags
+      delete currSavedTags[index]
+      logseq.updateSettings({
+        savedTags: 'Need to add some arbitrary string first',
+      })
+      logseq.updateSettings({ savedTags: currSavedTags })
+      logseq.hideMainUI()
     },
     [tags],
   )
@@ -97,33 +44,18 @@ export const ManageTags = ({
       <h2>Manage</h2>
       {Object.entries(localTags).map(([index, properties]) => (
         <div key={index} className="tag-management">
-          <h3>{index}</h3>
-          <DndContext sensors={sensors} onDragEnd={handleDragEnd}>
-            <SortableContext
-              items={properties.map((prop) => prop.name)}
-              strategy={verticalListSortingStrategy}
-            >
-              {properties.map(({ name }) => (
-                <SortableItem key={name} id={name} index={index}>
-                  {(attributes, listeners) => (
-                    <div className="sortable-property">
-                      <div
-                        className="icon-group"
-                        {...attributes}
-                        {...listeners}
-                      >
-                        <IconMenuOrder stroke={2} size="1rem" />
-                        <p>{name}</p>
-                      </div>
-                      <button onClick={() => deleteProperty(index, name)}>
-                        Delete
-                      </button>
-                    </div>
-                  )}
-                </SortableItem>
-              ))}
-            </SortableContext>
-          </DndContext>
+          <div className="tag-management-header">
+            <h3>{index}</h3>
+            <button onClick={() => deletePowertag(index)}>
+              Delete PowerTag
+            </button>
+          </div>
+          <TagProperties
+            setLocalTags={setLocalTags}
+            index={index}
+            properties={properties}
+            tags={tags}
+          />
         </div>
       ))}
     </div>
